requirements.txt
Django==1.4.2
MySQL-python==1.2.4
PIL==1.1.7
South==0.7.6
Werkzeug==0.8.3
django-bootstrap-toolkit==2.5.11
django-classy-tags==0.3.4.1
django-csv-importer==0.1.3.5
django-extensions==1.0.0
django-olwidget==0.60.3
django-sekizai==0.6.1
docutils==0.9.1
httplib2==0.7.6
ipython==0.13.1
psycopg2==2.4.6
pygraphviz==1.1
python-gflags==2.0
sorl-thumbnail==11.12
wsgiref==0.1.2
